chore(deps): bump apis-core to v0.15.1
apis-core deprecated the `name` field of the rootobject and changed it
to `deprecated_name` - we therefore introduce our own `name` fields.
This also ships a management command to copy values from
`deprecated_name` to `name`
apis-core also changed the approach of how the rdfimport module looks
for the correct rdf parser definition - we therefore update exising
definitions and add a couple of new ones.