chore(deps): bump apis-core to v0.23.0
The `GenericImporter` from apis-core was renamed to
`GenericModelImporter` and is thus picked automatically for all the
models inheriting from `GenericModel`. This means we don't have to
override it, if we don't want to change its functionality. It now also
automatically tries to import using RDF, therefore we can drop our
overrides for Place and Institution. For Person we still override the
importer, to be able to mangle that data before import, but at least the
inheriting `PersonImporter` now gets a bit smaller.