fix(autocomplete): set the _default_search_fields for Profession
Profession has multiple CharFields, including one that stores the name
of the Profession entry in the old APIS instance, based on which the
Profession instance was created. When looking up something in
autocomplete, by default all CharFields of a model are searched. This
leads to false positives when searching. Therefore we use the new and
shiny `_default_search_fields` to force the Profession model only to
search the `name` field.

Closes: #293